feat ✨: Update Dockerfile to use Python 3.10 and remove default installation
This commit updates the Dockerfile to use Python 3.10 instead of Python 3.11 for the base image, as Python 3.11 is no longer supported in some environments. Additionally, it removes the steps to install Python 3.11 from the PPA and set it as the default, which simplifies the Dockerfile and ensures compatibility with systems that do not support or prefer Python 3.11.